+++ src/dataStructures/disjointSet/__test__/disjointSet.test.ts
@@ -19,6 +19,7 @@ describe('DisjointSet', () => {
       .toBeNull();
 
     disjointSet.makeSet('A');
+    disjointSet.makeSet('A');
 
     expect(disjointSet.find('A'))
       .toBe('A');
@@ -68,12 +69,11 @@ describe('DisjointSet', () => {
     expect(disjointSet.inSameSet('A', 'C'))
       .toBe(true);
 
-    disjointSet
-      .makeSet('E')
-      .makeSet('F')
-      .makeSet('G')
-      .makeSet('H')
-      .makeSet('I');
+    disjointSet.makeSet('E');
+    disjointSet.makeSet('F');
+    disjointSet.makeSet('G');
+    disjointSet.makeSet('H');
+    disjointSet.makeSet('I');
 
     disjointSet.union('E', 'F');
     disjointSet.union('F', 'G');
@@ -96,10 +96,9 @@ describe('DisjointSet', () => {
   it('should union smaller set with bigger one making bigger one to be new root', () => {
     const disjointSet = new DisjointSet();
 
-    disjointSet
-      .makeSet('A')
-      .makeSet('B')
-      .makeSet('C');
+    disjointSet.makeSet('A');
+    disjointSet.makeSet('B');
+    disjointSet.makeSet('C');
     disjointSet.union('B', 'C');
     disjointSet.union('A', 'C');