+++ src/dataStructures/disjointSet/index.ts
@@ -1,3 +1,2 @@
 export * from './disjointSet';
 export * from './disjointSetItem';
-export * from './types';